Add API key for use by Brave Alert Lib
- Added corresponding test
- Refactored `/test` directory to encourage better organization into
  unit vs integration tests and to a single file per function
