+++ server/Installation.js
@@ -1,5 +1,5 @@
 class Installation {
-  constructor(id, name, responderPhoneNumber, fallbackPhoneNumbers, incidentCategories, isActive, createdAt) {
+  constructor(id, name, responderPhoneNumber, fallbackPhoneNumbers, incidentCategories, isActive, createdAt, apiKey) {
     this.id = id
     this.name = name
     this.responderPhoneNumber = responderPhoneNumber
@@ -7,6 +7,7 @@ class Installation {
     this.incidentCategories = incidentCategories
     this.isActive = isActive
     this.createdAt = createdAt
+    this.apiKey = apiKey
   }
 }
